+++ DB_ProjetLocation/Programmability/Stored Procedures/Goods/Section/Category/CSP_DeleteCategory.sql
@@ -2,6 +2,11 @@
 	@CategoryId INT
 AS
 BEGIN
-	DELETE FROM Category
-	WHERE Category_Id = @CategoryId;
+	IF NOT EXISTS(SELECT Good_Id FROM Good WHERE Category_Id = @CategoryId)
+		BEGIN
+			DELETE FROM Category
+			WHERE Category_Id = @CategoryId;
+		END
+	ELSE
+		RAISERROR('Unable to delete category', 16, 1);
 END
